refactor(domain): remove remaining dependency on interface contracts
