+++ src-tauri/src/domain/mod.rs
@@ -3,6 +3,7 @@ mod client_kind;
 mod client_profile;
 mod mutation_action;
 mod resource_kind;
+mod resource_record;
 
 pub use client_adapter::{AdapterListResult, AdapterMutationResult, ClientAdapter};
 pub use client_kind::ClientKind;
@@ -11,3 +12,4 @@ pub use client_profile::{
 };
 pub use mutation_action::MutationAction;
 pub use resource_kind::ResourceKind;
+pub use resource_record::ResourceRecord;
