+++ src-tauri/src/interface/contracts/list.rs
@@ -1,6 +1,7 @@
 use serde::{Deserialize, Serialize};
 
 use super::common::{ClientKind, ResourceKind};
+pub use crate::domain::ResourceRecord;
 use crate::infra::security::redaction::redact_sensitive_text;
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
@@ -10,22 +11,6 @@ pub struct ListResourcesRequest {
     pub resource_kind: ResourceKind,
     #[serde(default)]
     pub enabled: Option<bool>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ResourceRecord {
-    pub id: String,
-    pub client: ClientKind,
-    pub display_name: String,
-    pub enabled: bool,
-    pub transport_kind: Option<String>,
-    pub transport_command: Option<String>,
-    pub transport_args: Option<Vec<String>>,
-    pub transport_url: Option<String>,
-    pub source_path: Option<String>,
-    pub description: Option<String>,
-    pub install_kind: Option<String>,
-    pub manifest_content: Option<String>,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
